Create apis.go to refactor API related common methods
